requirements.txt
absl-py==0.7.1
astor==0.8.0
bitmex-ws==0.3.1
certifi==2019.6.16
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
pip==19.2.2
protobuf==3.8.0
pydot==1.4.1
pyparsing==2.4.2
pyreadline==2.1
PyYAML==5.1.2
scipy==1.3.1
setuptools==41.0.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
websocket-client==0.46.0
Werkzeug==0.15.5
wheel==0.33.4
wincertstore==0.2
wrapt==1.11.2
